Banking system with database
The system allows the creation of new accounts, authenticate account holders, and also manage credits and debits. Money can be deposited or withdrawn from an account provided the account holder gives proper authentication. When an account is opened, the details of account holders are recorded in the system’s files which includes their names, unique ids, address, account type, and phone numbers. An account can be owned by a maximum of two individuals (joint accounts). The system stores a secret alphanumeric key for each account. When the user transacts with the bank, this secret key is first required before any other transaction. Once the key is entered and verified the details of the account holder with current balance is displayed. A maximum of three attempts are allowed for entering the secret key, after which a relevant message is shown, and the user must exit the system. Uni Bank provides three types of accounts:  Checking Account, Saving Account, and Business Account. A checking account provides interest rates of 3% per annum. A maximum of AED 5000 can be withdrawn in one day and there must always be a balance of AED 100 in the account. For the saving account the interest rate is 7% per annum. Only a maximum of AED 2000 can be withdrawn in one day and there must always be a balance of AED 1000 in the account. For a business account, additional information is required like business name, business registration number, and business owner name. The interest rate is 3% per annum. Only a maximum of AED 10000 can be withdrawn in one day and there must always be a balance of AED 2000 in the account.