Vendi Machine System
Each item in a vending machine has item ID, item names, manufacturer name, price, and the quantity available in a machine. The items are of three types, liquids, sweet snacks, and salty snacks. The app on the phone displays all items that are available. A customer is a registered user of the mobile app whose name, address, and mobile number are stored in the system. When the phone application starts, the user puts the ID of the vending machine and the app displays the items and quantity of each available item in the machine. The customer should be able to select and pay for multiple items during the same session. When a customer pays for all selected item, the application should reduce the quantity of the selected item and also update and display the total amount of sales. Once the sale is complete, the customer is shown a random and unique alphanumeric key on the phone. When the customer punches in the key on the physical machine, the purchased items are dispensed to the customer.